+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### HEAD
 
 ```
+* -ssh option (issue #11)
+* Private key file detection for ssh clients (issue #10)
+* Username detection for ssh clients (issue #6)
 ```
 
 ### 0.3.1 (2014-03-20)